[US-199] Refacto variables and reorganise file tree
