+++ frontend/src/app/components/time-remaining/TimeRemaining.tsx
@@ -1,44 +1,41 @@
 import { Box, Text } from '@chakra-ui/react'
 import styles from './TimeRemaining.module.scss'
 import InformationContainer from '@/app/components/information-container/InformationContainer'
-import { Timer } from './Timer'
-import { useState } from 'react'
+import { Timer } from './timer/Timer'
 import { CurrentLevelInfosModel } from '@/app/models/LevelInfos.model'
+import useTimeRemaining from "@/app/components/time-remaining/useTimeRemaining";
 
 export default function TimeRemaining({
-  levelInfos,
+  currentLevelInfos,
 }: {
-  levelInfos: CurrentLevelInfosModel
+  currentLevelInfos: CurrentLevelInfosModel
 }) {
-  const [time, setTime] = useState(3605)
 
-  const handleTimeUp = () => {
-    setTime(time + 5)
-  }
+  const { time, handleTimeUp } = useTimeRemaining();
 
   return (
     <InformationContainer>
       <Box className={styles.container}>
         <Text>Temps restant : </Text>
         <Timer initialTime={time} onTimeUp={handleTimeUp} />
-        <Box className={styles.TimerZone}>
-          <Box className={styles.InfoZone}>
-            <Text className={styles.InfoText}>
+        <Box>
+          <Box className={styles.infoZone}>
+            <Text className={styles.infoText}>
               small blind / big blind / ante :
             </Text>
             <Text>
-              {levelInfos.smallBlindValue} /{levelInfos.bingBlindValue} /
-              {levelInfos.anteValue}
+              {currentLevelInfos.smallBlindValue} /{currentLevelInfos.bingBlindValue} /
+              {currentLevelInfos.anteValue}
             </Text>
           </Box>
-          <Box className={styles.InfoZone}>
-            <Text className={styles.InfoText}>BTN : </Text>
-            <Text>{levelInfos.btn}</Text>
+          <Box className={styles.infoZone}>
+            <Text className={styles.infoText}>BTN : </Text>
+            <Text>{currentLevelInfos.btn}</Text>
           </Box>
         </Box>
-        <Box className={styles.PotTotal}>
+        <Box className={styles.potTotal}>
           <Text>Pot Total :</Text>
-          <Text className={styles.Total}> {levelInfos.totalPot}</Text>
+          <Text className={styles.total}> {currentLevelInfos.totalPot}</Text>
         </Box>
       </Box>
     </InformationContainer>
